bigquery-mcp/bigquery-mcp-studio/pyproject.toml
[project]
name = "bigquery-mcp-studio"
version = "0.1.0"
description = "GUI for creating and maintaining parameterized BigQuery SQL queries for MCP"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-cloud-datastore>=2.19.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-auth>=2.35.0",
    "pydantic>=2.9.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
